service: flatten aws controller
