+++ service/controller/aws/error.go
@@ -1,8 +1,6 @@
 package aws
 
-import (
-	"github.com/giantswarm/microerror"
-)
+import "github.com/giantswarm/microerror"
 
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
