+++ service/controller/aws/legacy_cluster.go
@@ -17,7 +17,6 @@ import (
 	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	"k8s.io/client-go/kubernetes"
 
-	v22 "github.com/giantswarm/cluster-operator/service/controller/aws/v22"
 	"github.com/giantswarm/cluster-operator/service/internal/cluster"
 )
 
@@ -85,9 +84,9 @@ func NewLegacyCluster(config LegacyClusterConfig) (*LegacyCluster, error) {
 		}
 	}
 
-	var v22ResourceSet *controller.ResourceSet
+	var resourceSet *controller.ResourceSet
 	{
-		c := v22.ResourceSetConfig{
+		c := resourceSetConfig{
 			ApprClient:        config.ApprClient,
 			BaseClusterConfig: config.BaseClusterConfig,
 			CertSearcher:      config.CertSearcher,
@@ -106,7 +105,7 @@ func NewLegacyCluster(config LegacyClusterConfig) (*LegacyCluster, error) {
 			ResourceNamespace:  config.ResourceNamespace,
 		}
 
-		v22ResourceSet, err = v22.NewResourceSet(c)
+		resourceSet, err = newResourceSet(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -120,7 +119,7 @@ func NewLegacyCluster(config LegacyClusterConfig) (*LegacyCluster, error) {
 			Informer:  newInformer,
 			Logger:    config.Logger,
 			ResourceSets: []*controller.ResourceSet{
-				v22ResourceSet,
+				resourceSet,
 			},
 			RESTClient: config.G8sClient.CoreV1alpha1().RESTClient(),
