+++ service/controller/internal/chartconfig/chartconfig.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/micrologger"
 
 	"github.com/giantswarm/cluster-operator/pkg/annotation"
-	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/v22/key"
+	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/key"
 	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
 	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v22/key"
